+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [1.5.0]
 
+- loading on meals
+- order meals
 - add evolution screen
 - split personal nutriscore screen